+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/Security/Factory/OidcLoginFactory.php
@@ -34,6 +34,11 @@ class OidcLoginFactory extends AbstractFactory
 {
     public const PRIORITY = -25;
 
+    public function __construct()
+    {
+        $this->addOption('user_identifier_claim', 'sub');
+    }
+
     /**
      * @psalm-suppress ParamNameMismatch
      */
@@ -114,6 +119,11 @@ class OidcLoginFactory extends AbstractFactory
                 ->defaultValue([])
                 ->info('Additional parameters to include in the authorization request (e.g. prompt, max_age, display, ui_locales, acr_values, login_hint).')
             ->end()
+            ->enumNode('user_data_source')
+                ->values(['userinfo', 'id_token'])
+                ->defaultValue('userinfo')
+                ->info('Source of user claims: "userinfo" fetches from the UserInfo endpoint, "id_token" decodes claims from the ID token.')
+            ->end()
         ;
     }
 
@@ -171,6 +181,7 @@ class OidcLoginFactory extends AbstractFactory
         $options = array_intersect_key($config, $this->options);
         $options['firewall_name'] = $firewallName;
         $options['scope'] = $config['scope'];
+        $options['user_data_source'] = $config['user_data_source'];
         $options['pkce_enabled'] = $config['pkce']['enabled'];
         $options['pkce_method'] = $config['pkce']['method'];
         if (isset($config['max_age'])) {